Refactor PRIWA preview modules
