+++ .github/workflows/frontend-hosting-pull-request.yml
@@ -42,6 +42,8 @@ jobs:
       VITE_GEOPIFY_KEY: ${{ secrets.VITE_GEOPIFY_KEY }}
       VITE_SUPABASE_SENTINEL_PROCESSING_URL: ${{ secrets.VITE_SUPABASE_SENTINEL_PROCESSING_URL }}
       VITE_SUPABASE_SENTINEL_PROCESSING_ANON_KEY: ${{ secrets.VITE_SUPABASE_SENTINEL_PROCESSING_ANON_KEY }}
+      VITE_PRIWA_SUPABASE_URL: ${{ secrets.VITE_PRIWA_SUPABASE_URL }}
+      VITE_PRIWA_SUPABASE_ANON_KEY: ${{ secrets.VITE_PRIWA_SUPABASE_ANON_KEY }}
     steps:
       - uses: actions/checkout@v4
       - run: npm ci
